+++ tests/test_fitresult.py
@@ -12,24 +12,24 @@ true_b = 4.
 true_c = -0.3
 
 
-def create_loss():
+def create_loss(n=15000):
     a_param = zfit.Parameter("variable_a15151", 1.5, -1., 20.,
                              step_size=z.constant(0.1))
-    b_param = zfit.Parameter("variable_b15151", 3.5)
-    c_param = zfit.Parameter("variable_c15151", -0.04)
+    b_param = zfit.Parameter("variable_b15151", 3.5, 0, 20)
+    c_param = zfit.Parameter("variable_c15151", -0.04, -1, 0.)
     obs1 = zfit.Space(obs='obs1', limits=(-2.4, 9.1))
 
     # load params for sampling
-    a_param.load(true_a)
-    b_param.load(true_b)
-    c_param.load(true_c)
+    a_param.set_value(true_a)
+    b_param.set_value(true_b)
+    c_param.set_value(true_c)
 
     gauss1 = zfit.pdf.Gauss(mu=a_param, sigma=b_param, obs=obs1)
     exp1 = zfit.pdf.Exponential(lambda_=c_param, obs=obs1)
 
     sum_pdf1 = zfit.pdf.SumPDF((gauss1, exp1), 0.7)
 
-    sampled_data = sum_pdf1.create_sampler(n=15000)
+    sampled_data = sum_pdf1.create_sampler(n=n)
     sampled_data.resample()
 
     loss = zfit.loss.UnbinnedNLL(model=sum_pdf1, data=sampled_data)
@@ -124,3 +124,64 @@ def test_covariance(minimizer_class_and_kwargs):
     cov_mat_3_np = result.covariance(params=[a, b, c], method="hesse_np")
 
     np.testing.assert_allclose(cov_mat_3, cov_mat_3_np, rtol=0.05, atol=0.001)
+
+
+@pytest.mark.parametrize("minimizer_class_and_kwargs", minimizers)
+def test_errors(minimizer_class_and_kwargs):
+    results = create_fitresult(minimizer_class_and_kwargs=minimizer_class_and_kwargs)
+    result = results['result']
+    a = results['a_param']
+    b = results['b_param']
+    c = results['c_param']
+
+    z_errors, new_result = result.errors(method="zfit_error")
+    if new_result is not None:
+        result = new_result
+    minos_errors, _ = result.errors(method="minuit_minos")
+
+    for param in [a, b, c]:
+        z_error_param = z_errors[param]
+        minos_errors_param = minos_errors[param]
+        for dir in ["lower", "upper"]:
+            assert pytest.approx(z_error_param[dir], rel=0.03) == minos_errors_param[dir]
+
+
+@pytest.mark.xfail  # currently, fmin is not correct, loops, see: https://github.com/scikit-hep/iminuit/issues/395
+@pytest.mark.parametrize("minimizer_class_and_kwargs", minimizers)
+def test_new_minimum(minimizer_class_and_kwargs):
+    loss, params = create_loss(2000)
+
+    minimizer_class, minimizer_kwargs, test_error = minimizer_class_and_kwargs
+    minimizer = minimizer_class(**minimizer_kwargs)
+
+    if test_error:
+        ntoys = 100  # TODO: increase again?
+
+        new_minimum_found = False
+
+        for n in range(ntoys):
+            params_random_values = {p: np.random.uniform(p.lower_limit, p.upper_limit, ntoys) for p in params}
+
+            try:
+                zfit.param.set_values(params, [params_random_values[p][n] for p in params])
+                result_n = minimizer.minimize(loss=loss)
+                errors, new_result_n = result_n.errors(method='zfit_error')
+
+                if new_result_n is not None:
+                    new_minimum_found = True
+                    break
+
+            except RuntimeError:
+                continue
+
+        assert new_minimum_found
+
+        assert result_n.valid is False
+        for p in params:
+            assert errors[p] == "Invalid, a new minimum was found."
+
+        assert new_result_n.valid is True
+        errors, _ = new_result_n.errors()
+        for param in params:
+            assert errors[param]["lower"] < 0
+            assert errors[param]["upper"] > 0